requirements.txt
streamlit
psycopg2-binary
pandas
plotly
python-dotenv
fastapi
uvicorn
httpx
